Updated eof_varimax to allow missing values. Updated linint1/linint2 to allow input/output coordinates to be increasing or decreasing. Updated filwgts_lancos to allow "nsigma" to be a numeric.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6382 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/wrapper.c
@@ -4488,7 +4488,7 @@ void NclAddUserFuncs(void)
     SetArgTemplate(args,nargs,"integer",1,dimsizes);nargs++;
     SetArgTemplate(args,nargs,"numeric",1,NclANY);nargs++;
     SetArgTemplate(args,nargs,"numeric",1,NclANY);nargs++;
-    SetArgTemplate(args,nargs,"integer",1,dimsizes);nargs++;
+    SetArgTemplate(args,nargs,"numeric",1,dimsizes);nargs++;
     NclRegisterFunc(filwgts_lancos_W,args,"filwgts_lancos",nargs);
 
 /*
